% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_process_results.R
\name{get_process_results}
\alias{get_process_results}
\title{Interact with a background process}
\usage{
get_process_results(rbg_proc)
}
\arguments{
\item{rbg_proc}{an r_process returned by cherry_models(..., background = TRUE)}
}
\value{
data.frame when the `r_process` is finished; `invisible(NULL)` along with message on progress
otherwise
}
\description{
This function retrieves the result of a background process if the process if finished computing.
In this case the handler (the 'process object') will be removed from the environment.
If it is still running the output of the process, such as messages, are printed.
}
\examples{
proc <- cherry_models(
  model_data = mpg,
  response = "cty",
  fixed_inter = c("displ", "cyl", "year"),
  fixed_no_inter = c("hwy"),
  max_comb = 3,
  comb_comb = 3,
  random = "(1|model)",
  n.workers = 2,
  dry = FALSE,
  background = TRUE
)

get_process_results(proc)

}
