% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_report.R
\name{show_report}
\alias{show_report}
\title{Display a report in a shiny app}
\usage{
show_report(data)
}
\arguments{
\item{data}{data.frame returned by `cherry_models()` or `get_process_results()`}
}
\value{
starts a shiny app
}
\description{
Click on the `i` button in the top right corner for help.
}
\examples{

x <- cherry_models(
  model_data = mpg,
  response = "cty",
  fixed_inter = c("displ", "cyl", "year"),
  fixed_no_inter = "hwy",
  max_comb = 3,
  comb_comb = 2,
  random = "(1|model)",
  n.workers = 2,
  dry = FALSE,
  background = FALSE
)

show_report(x)

}
